# Improve worst-case NameAbbreviator performance